+++ controller/editorcontroller.php
@@ -35,12 +35,15 @@ use OCP\IL10N;
 use OCP\ILogger;
 use OCP\IRequest;
 use OCP\ISession;
+use OCP\ITagManager;
 use OCP\IURLGenerator;
 use OCP\IUser;
 use OCP\IUserSession;
 use OCP\Share\IManager;
 
 use OCA\Files\Helper;
+
+use OC\Tags;
 
 use OCA\Onlyoffice\AppConfig;
 use OCA\Onlyoffice\Crypt;
@@ -119,6 +122,13 @@ class EditorController extends Controller {
     private $versionManager;
 
     /**
+     * Tag manager
+     *
+     * @var ITagManager
+    */
+    private $tagManager;
+
+    /**
      * Mobile regex from https://github.com/ONLYOFFICE/CommunityServer/blob/v9.1.1/web/studio/ASC.Web.Studio/web.appsettings.config#L35
      */
     const USER_AGENT_MOBILE = "/android|avantgo|playbook|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od|ad)|iris|kindle|lge |maemo|midp|mmp|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\\/|plucker|pocket|psp|symbian|treo|up\\.(browser|link)|vodafone|wap|windows (ce|phone)|xda|xiino/i";
@@ -135,6 +145,7 @@ class EditorController extends Controller {
      * @param Crypt $crypt - hash generator
      * @param IManager $shareManager - Share manager
      * @param ISession $ISession - Session
+     * @param ITagManager $tagManager - Tag manager
      */
     public function __construct($AppName,
                                     IRequest $request,
@@ -146,7 +157,8 @@ class EditorController extends Controller {
                                     AppConfig $config,
                                     Crypt $crypt,
                                     IManager $shareManager,
-                                    ISession $session
+                                    ISession $session,
+                                    ITagManager $tagManager
                                     ) {
         parent::__construct($AppName, $request);
 
@@ -157,6 +169,7 @@ class EditorController extends Controller {
         $this->logger = $logger;
         $this->config = $config;
         $this->crypt = $crypt;
+        $this->tagManager = $tagManager;
 
         $this->versionManager = new VersionManager($AppName, $root);
 
@@ -1114,6 +1127,9 @@ class EditorController extends Controller {
             $createUrl = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".editor.create_new", $createParam);
 
             $params["editorConfig"]["createUrl"] = urldecode($createUrl);
+
+            $params["document"]["info"]["favorite"] = $this->isFavorite($fileId);
+            $params["_file_path"] = $userFolder->getRelativePath($file->getPath());
         }
 
         if ($folderLink !== null
@@ -1349,6 +1365,22 @@ class EditorController extends Controller {
         }
 
         return $params;
+    }
+
+    /**
+     * Check file favorite
+     *
+     * @param integer $fileId - file identifier
+     *
+     * @return bool
+     */
+    private function isFavorite($fileId) {
+        $currentTags = $this->tagManager->load("files")->getTagsForObjects([$fileId]);
+        if ($currentTags) {
+            return in_array(Tags::TAG_FAVORITE, $currentTags[$fileId]);
+        }
+
+        return false;
     }
 
     /**